Revert "moar extractions"
This reverts commit 4da04ec383be552abe8595e820475cca048fcb49.
