+++ Explorer/Assets/DCL/Chat/ChatController.cs
@@ -154,8 +154,7 @@ namespace DCL.Chat
             conversationEventsHelper = new ChatControllerConversationEventsHelper(
                 chatHistory,
                 chatUserStateUpdater,
-                this,
-                chatStorage);
+                this);
 
             userStateHelper = new ChatControllerUserStateHelper(
                 chatUserStateUpdater,
@@ -165,8 +164,7 @@ namespace DCL.Chat
                 chatHistory,
                 this,
                 chatBubblesHelper,
-                chatSettings,
-                hyperlinkTextFormatter);
+                chatSettings);
         }
 
 #region Panel Visibility
@@ -378,12 +376,18 @@ namespace DCL.Chat
 #region Channel Events
         private async void OnViewCurrentChannelChangedAsync()
         {
-            await conversationEventsHelper.OnCurrentChannelChangedAsync();
+            if (chatHistory.Channels[viewInstance!.CurrentChannelId].ChannelType == ChatChannel.ChatChannelType.USER &&
+                chatStorage != null && !chatStorage.IsChannelInitialized(viewInstance.CurrentChannelId))
+            {
+                await chatStorage.InitializeChannelWithMessagesAsync(viewInstance.CurrentChannelId);
+                chatHistory.Channels[viewInstance.CurrentChannelId].MarkAllMessagesAsRead();
+                viewInstance.RefreshMessages();
+            }
         }
 
         private void OnViewChannelRemovalRequested(ChatChannel.ChannelId channelId)
         {
-            conversationEventsHelper.OnChannelRemovalRequested(channelId);
+            chatHistory.RemoveChannel(channelId);
         }
 #endregion
 
@@ -460,12 +464,24 @@ namespace DCL.Chat
 
         private void OnChatBusMessageAdded(ChatChannel.ChannelId channelId, ChatMessage chatMessage)
         {
-            messageHandlingHelper.OnChatBusMessageAdded(channelId, chatMessage);
+            if (!chatMessage.IsSystemMessage)
+            {
+                string formattedText = hyperlinkTextFormatter.FormatText(chatMessage.Message);
+                var newChatMessage = ChatMessage.CopyWithNewMessage(formattedText, chatMessage);
+                chatHistory.AddMessage(channelId, newChatMessage);
+            }
+            else
+                chatHistory.AddMessage(channelId, chatMessage);
         }
-
         private void OnViewMemberListVisibilityChanged(bool isVisible)
         {
-            memberListHelper.OnViewMemberListVisibilityChanged(isVisible);
+            if (isVisible && roomHub.HasAnyRoomConnected())
+                RefreshMemberList();
+        }
+
+        private void RefreshMemberList()
+        {
+            memberListHelper.RefreshMemberList();
         }
 
         private void GetProfilesFromParticipants(List<Profile> outProfiles)
