+++ Explorer/Assets/DCL/Chat/ChatControllerMemberListHelper.cs
@@ -125,11 +125,5 @@ namespace DCL.Chat
         {
             memberListCts.SafeCancelAndDispose();
         }
-
-        public void OnViewMemberListVisibilityChanged(bool isVisible)
-        {
-            if (isVisible && roomHub.HasAnyRoomConnected())
-                RefreshMemberList();
-        }
     }
 }
